Move Git Credential Store setup to initPlatform()
