+++ lib/modules/platform/gitlab/index.ts
@@ -1,8 +1,11 @@
 import URL from 'node:url';
+import { homedir } from 'os';
 import { setTimeout } from 'timers/promises';
 import is from '@sindresorhus/is';
+import fs from 'fs-extra';
 import pMap from 'p-map';
 import semver from 'semver';
+import upath from 'upath';
 import { GlobalConfig } from '../../../config/global';
 import {
   CONFIG_GIT_URL_UNAVAILABLE,
@@ -127,6 +130,22 @@ export async function initPlatform({
   const platformConfig: PlatformResult = {
     endpoint: defaults.endpoint,
   };
+
+  if (GlobalConfig.get('gitCredentialPassing') === 'store') {
+    logger.debug('Enable the Git credential store');
+    await git.enableCredentialStore();
+
+    const lines: string[] = [];
+    for (const host of hostRules.getAll() ?? []) {
+      lines.push(`https://oauth2:${host.token}@${host.matchHost}`);
+    }
+    const credentialsPath = upath.join(homedir(), '.git-credentials');
+    await fs.writeFile(credentialsPath, lines.join('\n') + '\n', {
+      mode: 0o600,
+      encoding: 'utf-8',
+    });
+  }
+
   let gitlabVersion: string;
   try {
     if (!gitAuthor) {
